clean aides.jsx
